major cleanup
realized that github should really be hosting the XML files and the ReST
results, not the entire dev environment, which should be hosted in
parent directory. The more you learn I guess.

Included also the original Blogger XML file, and the pre-cleaned
WordPress XML converted file (via
https://blogger2wordpress.appspot.com/). Might as well test it out again
if the pelican issue with invalid filenames gets fixed.